Added notype channels into an array instead of doing individual checks

+++ discord/events/messageCreate.js
@@ -10,13 +10,14 @@ module.exports = {
   name: "messageCreate",
   async execute(message) {
     if (message.author.bot) return;
-    if (
-      message.channel.id === REGISTRATION_ID ||
-      message.channel.id === TICKETS_ID ||
-      message.channel.id === AWARD_VOTING_ID
-    ) {
-      await message.delete();
-    }
+
+    const NO_TYPE_CHANNEL_IDS = [REGISTRATION_ID, TICKETS_ID, AWARD_VOTING_ID];
+
+    const isInNoTypeChannel = NO_TYPE_CHANNEL_IDS.some(
+      (channelID) => channelID === message.channel.id
+    );
+
+    if (isInNoTypeChannel) await message.delete();
 
     if (message.channel.id === SUGGESTIONS_ID) {
       message.react(OkHand);
